// src/api/models/mCQData.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Multiple Choice Question data.
 */
export interface MCQData {
  /**
   * @minItems 2
   * @maxItems 10
   */
  options: string[];
  /** @minimum 0 */
  correct_index: number;
}