// src/api/models/mRQData.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Multiple Response Question data.
 */
export interface MRQData {
  /**
   * @minItems 2
   * @maxItems 10
   */
  options: string[];
  /** @minItems 1 */
  correct_indices: number[];
}