// src/api/models/shortAnswerData.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Short Answer Question data.
 */
export interface ShortAnswerData {
  /**
   * @minLength 1
   * @maxLength 5000
   */
  correct_answer: string;
}