// src/api/models/question.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionQuestionType } from './questionQuestionType';

/**
 * Represents a generated question.
 */
export interface Question {
  /** The text of the question */
  text: string;
  /** The answer to the question */
  answer: unknown;
  /** List of options for MCQ or MRQ types */
  options?: unknown[] | null;
  /** List of indices of correct options for MCQ or MRQ types */
  correct_indices?: number[] | null;
  /** Type of the question (mcq, mrq, short_answer, etc.) */
  question_type: QuestionQuestionType;
}